requirements.txt
numpy==2.4.4
pandas==3.0.2
Pillow==9.0.1
Pillow==12.2.0
plotly==6.7.0
psutil==7.2.2
python-dotenv==1.2.2
Requests==2.33.1
scapy==2.7.0
streamlit==1.57.0
